No longer mangle errors from unmarshalResponse
Juju is calling the `Do` method which calls unmarshalResponse. This library prepends things to the returned error messages which creates an impossible situation for callers. They cannot retrieve a clean version of the errors the server returns without importing errgo and introspecting on the error value.

This patch removes the annotations unless there has been an error during unmarshaling.