rename ErrorMapper to Server; remove UUID support
Also change the handlers function type so that it returns a context, so
that it can create a custom context for its handler methods, pass the
context to the error mapper and remove
direct support for request UUIDs - we now have the capability to support
them externally.  There are many possible formats for request-associated
context and httprequest doesn't seem like the right place to make that
decision.

Note that this is a backwardly incompatible change.