fix handling of anonymous fields
cherry picked from 250d073dcf1e237b3bf9849cfa6cb56a715b77de

When an anonymous body-tagged field was unmarshaled,
all its fields were considered too, and allocated if they
were pointers.

This PR fixes that so that all fields within tagged anonymous
fields are ignored for unmarshaling purposes,
which also addresses a previous TODO in the code.

As part of debugging this, we added a name field to the field
type, which made it trivial to fix another TODO to add the
field name to the unmarshal error message.